# test_jit.py: Speedup EndToEnd tests by reducing workload size.