@ZolotukhinM 2019-01-10
With these changes the tests take around 10s each, which is still a lot to me (most of other tests take fractions of a second), but at least we don't suffer as much as before.